Update patch set 2
Patch Set 2:

I wonder will the API component have an independent life as a separate component or simply be a Java library that other components will use. If not, then it may not need start stop, command line arguments, and activation.

Also on parameters, it probably won't need a separate parameter handler as that type of functionality will be in the executable component that calls the library.

Patch-set: 2