Update patch set 1
Patch Set 1: Code-Review-1

For the Drools (which is located in drools-applications), the testsuite has a pom.xml and subsequent pom.xml for stability and performance. Apex has this. I understand policy/engine didn't have this and neither did distribution, but I'd like for there to be consistency between all the repos as to where the tests are and how to run them.

Patch-set: 1
Label: Code-Review=-1