Update patch set 2
Patch Set 2:

I would actually like to check with Sebastien, if I could make a slight change to the Operational policy request.
Can I change the input to be similar to casablanca, but the following:
{
"policy-id": "scale.operational",
"contents" : "<URL ESCAPED YAML>"
}

instead of just the Unescaped YAML. Eg make it similar to the guard policies. The problem is we should allow CLAMP to specify a name for operational and guard, rather than inferring something. There could be the possibility of clashing.

Patch-set: 2
Label: Code-Review=0