Update patch set 1
Patch Set 1:

With regards to error messages, it would be best to collect the info into an ErrorDetails object or something of that sort.   An object that can be automatically converted into json (or yaml) by the jax-rs middleware, rather than composing a string with newlines, etc .. (which may also not fit the content-type to be sent).

For the error message, I think we want to send meaningful error description, but not stack traces, etc .. it may be good for our own debugging but of no value to external API clients to make sense of it.   It is also not a recommended practice from a security's standpoint to leak internals of the service like the stacktrace.

I suggest to create an "Error" object for such that can be serialized automatically by middleware into the right format, and not include stack trace information, all of that should be able to be found in the logs.

Patch-set: 1