Update patch set 2
Patch Set 2:

Is preloading policy types a tentative thing in Dublin and will be deprecated in next release or we would like it to be a long-term scheme? If tentative, I'd like to do it this way so that we can easily deprecate ApiDatabaseInitializer in El Alto without changing ParameterGroup and bunches of config files. If long-term, better to put it in config.

Patch-set: 2