Update patch set 2
Patch Set 2:

> (1 comment)

We start checking the existence of policy type for legacy policies. Although we have them pre-loaded, we use h2 in-mem db for junit tests and config it to be drop-and-create-tables. Hence, each new db connection will cause previous tables to drop. Pre-loaded policy types cannot be persistent in junit.

Patch-set: 2