Update patch set 1
Patch Set 1:

> (1 comment)

I thought pdpState=ACTIVE is equivalent to "at least one instance of pdp is registered to this group". If no pdp is registered, then it is not possible pdpState=ACTIVE (this means at least one pdp in this group is ACTIVE if I understand correctly). Maybe we can keep it?

Patch-set: 1