Update patch set 1
Patch Set 1:

OK, from integration testing, the most effective is the major.minor version, ie. 1.4-SNAPSHOT-latest, when working in a release.    This avoids having to change OOM, integration repos, and helm charts when working with integration teams in labs, which is more disruptive.

Taking a look at nexus 3 that somehow, "released versions" (or at least versioned as releases) are in docker snapshot that don't match the same released ones in docker releases.   For example 2.0.0. 2.0.1, and 2.0.2 (for dublin dev. branch), that may cause problems ..

Patch-set: 1