Update patch set 3
Patch Set 3:

> (1 comment)

We could in a different ticket as for consistency we would want all Dockerfiles to change.  It also requires to change the FROM and move to snapshot image.   As a side note, I think the ARG+ENV combo in parent Dockerfile is the one that causes problems, BUILD_VERSION I think would (which is what I observed which has ARG+ENV)  but POLICY_HOME (only uses ENV).

Patch-set: 3