Update patch set 1
Patch Set 1: Code-Review+1

I'm good with this as is - these messages are annoying. But I'm wondering if we should pull into a configuration file?? Or use logback.xml to remove them?? Hardcoding in the JUnits is one thing, but for production maybe there's a remote chance they need to be put back on for debugging??. Perhaps its ok and too much work for little return, what do you folks think?

Patch-set: 1
Label: Code-Review=+1