Update patch set 2
Patch Set 2:

(1 comment)

> > (1 comment)
 > 
 > wH
 > 
 > > (1 comment)
 > 
 > I am not sure I get what you mean. Could you expand a little bit?
 > Jorge

The main question I had

1. Is the "deployed" API identical in the format of the response including return response to the proper tosca one? 

The reason for this question is to know if it is fully transparent to the way that CLAMP coded it up.   If not, we need to see and if divergence is needed from the proper tosca policies we need to feed that information back to the CLAMP team.

The other question that I was mulling may be OK, now thinking more about it.   This API can return a list of multiple policies to the caller and that may be OK as in theory the same policy may be deployed across multiple groups.   As of now though, we only support one group, so there will only 0, or 1 deployed, here the question I was entertaining if a list is appropriate for the caller.   No question at this point.  But let me know if you guys have any thoughts in the matter.

Patch-set: 2