Update patch set 4
Patch Set 4:

(1 comment)

We just need to confirm, that once this is merged we can create a policy via the API, deploy it the xacml pdp, ensure that the guard translates it appropriate, and ultimately a Decision API takes the CLC policy into account. Sounds like a lot, but this needs to be done else this work is sort of useless.

Patch-set: 4