Update patch set 10
Patch Set 10:

Just wondering are these policy types that we will have as definitive policy types in the system. If so, they probably should be stored in policy-models with all the other definitive policy types.

Here they are being stored in the test directory.

If you want to use them here for testing, you can use the policy models jar as a resource (its already a dependency of policy-api) and get at them that way.

Patch-set: 10