Update patch set 10
Patch Set 10: Code-Review+2

Ok so via email, Avteet indicated to leave the identity field in. So if we can please submit this so I can copy it to the appropriate places and move on to the code that would be helpful.

Patch-set: 10
Reviewer: Gerrit User 57 <57@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+2