Knock on of changing policy types to map
This review adapts policy-api for the change in the structure
of the toasca service template from a list of maps to a plain map.

There is a lot of duplication in the policy type definitions, which
should be pulled from policy-models. That work will be done in a future
review. In this review, the structure of the existing test data was
fixed, sometimes more than once for the same policy type definition.

Issue-ID: POLICY-2047
Change-Id: Ie1fdeef79289348005314c5de9b2e37c0c45d786
Signed-off-by: liamfallon <liam.fallon@est.tech>