Update patch set 2
Patch Set 2:

(7 comments)

All the preloaded policies for optimization need to be redone - they are still the Dublin versions. Probably the best thing to do is convert the YAML in src/test/resources to JSON and overwrite the old policies. But first make sure we quote the operators appropriately.

Patch-set: 2