Update patch set 4
Patch Set 4:

> (1 comment)

I am ok to reuse your validation framework per policy type. But it is better to be in models or common. It looks a bit strange for api to refer to a class that is in pdp repo. Perhaps we could leave these validation as they are at the moment. Then in G release, we can consolidate all kinds of validation under the same umbrella (one storey/epic could be raised for that).

Patch-set: 4
Label: Code-Review=0