Update patch set 1
Patch Set 1:

Just looking at the API code, the only place where there is a problem with the database not coming up is in the initiation of the policy types and policies during API startup.

In the class org.onap.policy.api.main.startstop.Main at line 75, the code tries to read the predefined policy types and policies from the DB and if the DB read fails, it crashes out.

However, the DB calls coming off the API in the ApiRestController do not bring down the JVM if they fail,t hey simply return a HTTP error code INTERNAL_SERVER_ERROR (500) but continues running.

So you only need to take care of the case at line 75 in Main. You can use a Sleep here or just spin off a thread to do the DB read in a loop and wait for it to return.

Once the DB is read int he startup you're good, the current code is fine.

Patch-set: 1
Reviewer: Gerrit User 2787 <2787@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>