Update patch set 3
Patch Set 3:

(13 comments)

Still outstanding for discussion:
Do we keep this code here (in API and in PAP) or push it to policy.models?
What test do we use to determine whether the connection should be retried? How do we parameterize the test if necessary?

Also need to decide how to parameterize delayPeriod

Patch-set: 3