Update patch set 1
Patch Set 1:

> Ahh! you are correct, the one that ran was master verify. I remember we use to have it earlier. Did we stopped keeping them?

Unfortunately, with the restructuring of the CSITs, it became more difficult to do a verify, as the master script in docker/csit only pulls the head of the branch for the repo under test.  Furthermore, the test requires a docker image of the repo to have been built and pushed to dockerio/nexus.

Because of the above, I'm wondering at the value of having each repo's CSIT kept within the repo instead of in the docker/csit.  It was thought that we'd update the CSITs at the same time as making other changes to the repo, but not sure that's really all that likely.  In addition, because it also causes all of the above issues, maybe we should move it all back to docker/csit, then we could reinstitute CSIT verifies, as well.

Patch-set: 1