Update patch set 1
Patch Set 1:

The issue here is that preloading of policy types when there are multiple versions only happen for one of them (the latest as it goes in sequence loading them).    The code added here will detect if another version of the same policy type is preloaded and if so, it will be preloaded separately instead of in the final merged service template as the code currently does.

Patch-set: 1