Use generated API interface for Swagger
This commit removes the Swagger V2 annotations on the
NodeTemplateController in policy-api. The OpeApi annotations (Swagger
v3 annotations) are on a generated Java Interface. The code is changed
so that the controller implements that interface.

There are no code changes tot he controller except that the order of the
parameters is switched in some cases.

Issue-ID: POLICY-4404
Change-Id: Iab49c7f3f197a85d3314a879d39fe2c95be041b8
Signed-off-by: liamfallon <liam.fallon@est.tech>