Use generated API interface for Swagger(2)
This commit removes the Swagger V2 annotations on the
ApiRestController in policy-api. The OpeApi annotations (Swagger
v3 annotations) are on a generated Java Interface. The code is changed
so that the controller implements that interface.

There are no code changes tot he controller except that the order of the
parameters is switched in some cases.

Issue-ID: POLICY-4404
Change-Id: Ic8976362481a0b49ad05cebb2f4f9ce4147f3965
Signed-off-by: lapentafd <francesco.lapenta@est.tech>